+++ app/components/Icon/Icon.jsx
@@ -9,7 +9,7 @@ let icons = ["user", "trash", "chevron-down", "menu", "database", "download", "s
     "thumb-tack", "clock", "clippy", "shuffle", "transfer", "dollar", "deposit", "withdraw",
     "settle", "trade", "adjust", "excel", "share", "minus-circle", "cogs", "dashboard",
     "server", "power", "thumbs-up", "folder", "warning", "gift", "text", "list",
-    "hourglass"];
+    "hourglass", "news"];
 
 let icons_map = {};
 for (let i of icons) icons_map[i] = require(`./${i}.svg`);
@@ -25,13 +25,13 @@ class Icon extends React.Component {
         if(this.props.className) {
             classes += " " + this.props.className;
         }
-        return <span className={classes} style={this.props.style || {}} dangerouslySetInnerHTML={{__html: icons_map[this.props.name]}}/>;
+        return <span className={classes} style={this.props.style || {}} dangerouslySetInnerHTML={{__html: icons_map[this.props.name]}} onClick={this.props.onClick} />;
     }
 }
 
 Icon.propTypes = {
     name: React.PropTypes.string.isRequired,
-    size: React.PropTypes.oneOf(["1x", "2x", "3x", "4x", "5x", "10x"]),
+    size: React.PropTypes.oneOf(["1x", "1_5x", "2x", "3x", "4x", "5x", "10x"]),
     inverse: React.PropTypes.bool,
     className: React.PropTypes.string
 };