+++ app/electron_imports.js
@@ -41,3 +41,4 @@ import Brainkey from "./components/Wallet/Brainkey";
 import Help from "./components/Help";
 import InitError from "./components/InitError";
 import CreateWorker from "./components/Account/CreateWorker";
+import News from "./components/News";