+++ app/stores/SettingsStore.js
@@ -86,7 +86,7 @@ class SettingsStore {
             themes: [
                 "darkTheme",
                 "lightTheme",
-                "olDarkTheme"
+                "midnightTheme"
             ],
             passwordLogin: [
                 {translate: "cloud_login"},
@@ -99,12 +99,18 @@ class SettingsStore {
         };
 
         this.settings = Immutable.Map(merge(this.defaultSettings.toJS(), ss.get("settings_v3")));
-
+        if (this.settings.get("themes") === "olDarkTheme") {
+            this.settings = this.settings.set("themes", "midnightTheme");
+        }
         let savedDefaults = ss.get("defaults_v1", {});
         /* Fix for old clients after changing cn to zh */
         if (savedDefaults && savedDefaults.locale) {
             let cnIdx = savedDefaults.locale.findIndex(a => a === "cn");
             if (cnIdx !== -1) savedDefaults.locale[cnIdx] = "zh";
+        }
+        if (savedDefaults && savedDefaults.themes) {
+            let olIdx = savedDefaults.themes.findIndex(a => a === "olDarkTheme");
+            if (olIdx !== -1) savedDefaults.themes[olIdx] = "midnightTheme";
         }
         if (savedDefaults.apiServer) {
             savedDefaults.apiServer = savedDefaults.apiServer.filter(a => {
@@ -171,7 +177,8 @@ class SettingsStore {
                     "OPEN.USDT", "OPEN.EURT", "OPEN.BTC", "CADASTRAL", "BLOCKPAY", "BTWTY",
                     "OPEN.INCNT", "KAPITAL", "OPEN.MAID", "OPEN.SBD", "OPEN.GRC",
                     "YOYOW", "HERO", "RUBLE", "SMOKE", "STEALTH", "BRIDGE.BCO",
-                    "BRIDGE.BTC", "KEXCOIN", "PPY", "OPEN.EOS", "OPEN.OMG"
+                    "BRIDGE.BTC", "KEXCOIN", "PPY", "OPEN.EOS", "OPEN.OMG", "CVCOIN",
+                    "BRIDGE.ZNY", "BRIDGE.MONA", "OPEN.LTC"
                 ],
                 markets_39f5e2ed: [ // TESTNET
                     "PEG.FAKEUSD", "BTWTY"